refactor: extract linear combination of Columns
To prepare additional changes later.  But it's easier to review in two
parts.

Extracted from https://github.com/0xmozak/mozak-vm/pull/472
